Replace base and remove traces of old base
