+++ GPTifier/src/networking/api_openai_user.hpp
@@ -1,11 +1,8 @@
 #pragma once
 
-#include "curl_base.hpp"
-
 #include <string>
 
 namespace networking {
-
 std::string get_models();
 std::string delete_model(const std::string &model_id);
 std::string create_chat_completion(const std::string &post_fields);
@@ -17,11 +14,4 @@ std::string get_uploaded_files(bool sort_asc = true);
 std::string delete_file(const std::string &file_id);
 std::string create_fine_tuning_job(const std::string &post_fields);
 std::string get_fine_tuning_jobs(const std::string &limit);
-
-class OpenAIUser: public CurlBase {
-public:
-private:
-    void reset_handle();
-};
-
 } // namespace networking
